refactor: 점수를 계산하는 기능에 대한 메서드 분리

+++ src/test/java/chess/board/ChessBoardTest.java
@@ -583,7 +583,7 @@ class ChessBoardTest {
 
         ChessBoard chessBoard = ChessBoard.createBoardByRule(piecePosition);
         // when
-        List<Integer> scores = chessBoard.calculateScore();
+        List<Double> scores = chessBoard.calculateScore();
 
         // then
         assertThat(scores.get(0)).isEqualTo(18);
@@ -602,7 +602,7 @@ class ChessBoardTest {
 
         ChessBoard chessBoard = ChessBoard.createBoardByRule(piecePosition);
         // when
-        List<Integer> scores = chessBoard.calculateScore();
+        List<Double> scores = chessBoard.calculateScore();
 
         // then
         assertThat(scores.get(1)).isEqualTo(17);
